requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
discord.py==2.4.0
emoji==2.14.0
frozenlist==1.5.0
idna==3.10
loguru==0.7.2
multidict==6.1.0
opuslib==3.0.1
propcache==0.2.0
psutil==6.1.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
soupsieve==2.6
spotipy==2.24.0
urllib3==2.2.3
win32-setctime==1.1.0
yarl==1.17.1
youtube-dl==2021.12.17
yt-dlp==2024.11.4
